pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=physics_tools --cov-report=term-missing"

[project]
name = "physics-utils"
version = "0.3.0"
authors = [
  { name="Jahan Rashidi", email="hiimjahan@gmail.com" },
]
description = "Helpers for empirical data handling and plotting"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "GPL-3.0-only"
license-files = ["LICEN[CS]E*"]

dependencies = [
    "numpy~=2.3.1",
    "matplotlib~=3.10.6",
    "scipy~=1.16.2",
    "pandas~=2.3.2",
    "antlr4-python3-runtime~=4.13.0"
]

[tool.setuptools.packages.find]
exclude = ["tests", "docs", "examples", "labs"]

[project.urls]
Homepage = "https://github.com/ImNotJahan/PhysicsTools"
Source = "https://github.com/ImNotJahan/PhysicsTools"
Issues = "https://github.com/ImNotJahan/PhysicsTools/issues"
Documentation = "https://physics-utils.readthedocs.io"